ge: split include files between user & kernel
